+++ packages/app-react/src/components/Details/index.tsx
@@ -4,7 +4,7 @@ import { isEmpty, isFunction, isPlainObject, isString, mergeWith } from 'lodash-
 import { Fragment, useState } from 'react';
 
 import Card from '@/components/Card';
-import createOnClickCopyToClipboard from '@fieryeagle/browser-clipboard';
+import createOnClickCopyToClipboard from '@cheshirecode/browser-clipboard';
 import { isEmptyObject } from '@/utils';
 
 import type { DetailsData, DetailsMetadata, DetailsProps } from './typings';